(repository): ajout de l'implementation en dur de plusieurs recettes(recipes_repository, recipes_repository_dummy_impl)

+++ data/Ingredient.swift
@@ -8,6 +8,6 @@ import Foundation
 
 struct Ingredient{
     var name : String
-    var quantity : Int
+    var quantity : Float
     var unit : String
 }
